Fix bug in spawn mode of multiprocessing
