+++ src/agentscope/utils/__init__.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 """ Import modules in utils package."""
 from .monitor import MonitorBase
-from .monitor import QuotaExceededError
 from .monitor import MonitorFactory
 
 __all__ = [
     "MonitorBase",
-    "QuotaExceededError",
     "MonitorFactory",
 ]
