+++ tests/knowledge_test.py
@@ -9,7 +9,6 @@ from typing import Any
 import shutil
 
 import agentscope
-from agentscope.manager import FileManager
 from agentscope.rag import LlamaIndexKnowledge
 from agentscope.models import OpenAIEmbeddingWrapper, ModelResponse
 
@@ -46,7 +45,6 @@ class KnowledgeTest(unittest.TestCase):
 
     def tearDown(self) -> None:
         """Clean up before & after tests."""
-        FileManager.flush()
         try:
             if os.path.exists(self.data_dir):
                 shutil.rmtree(self.data_dir)
