+++ tests/service_toolkit_test.py
@@ -5,7 +5,6 @@ import unittest
 from typing import Literal
 
 import agentscope
-from agentscope.manager import FileManager
 from agentscope.models import ModelWrapperBase, ModelResponse
 from agentscope.parsers import MultiTaggedContentParser, TaggedContent
 from agentscope.service import (
@@ -378,10 +377,6 @@ The following tool functions are available in the format of
                 },
             },
         )
-
-    def tearDown(self) -> None:
-        """Clean up the test environment"""
-        FileManager.flush()
 
 
 if __name__ == "__main__":
